Renamed module convolutions to fft
- This is done to follow the structure of numpy
- Also, expose the functions through the init to be able to call `qttools.fft.`
- Change all files accordingly where functions were imported.
