--- src/qttools/convolutions/__init__.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2024-2026 ETH Zurich and the authors of the qttools package.
