+++ src/qttools/fft/__init__.py
@@ -1,0 +1,15 @@
+# Copyright (c) 2024-2026 ETH Zurich and the authors of the qttools package.
+
+from qttools.fft.ffts import (
+    fft_circular_convolve,
+    fft_convolve,
+    fft_convolve_kpoints,
+    fft_correlate_kpoints,
+)
+
+__all__ = [
+    "fft_convolve",
+    "fft_circular_convolve",
+    "fft_convolve_kpoints",
+    "fft_correlate_kpoints",
+]
