+++ tests/quatrex/electron/test_sse_coulomb_screening.py
@@ -3,7 +3,7 @@
 import pytest
 
 from qttools import NDArray, xp
-from qttools.convolutions.convolve import _naive_convolve
+from qttools.fft.convolve import _naive_convolve
 from quatrex.electron.sse_coulomb_screening import hilbert_transform
